general changes

+++ routers/user.js
@@ -7,7 +7,7 @@ const smsClient = require('twilio')(accountSid, authToken);
 const router = new express.Router();
 
 //Create the user in the database and send an sms code back
-router.post('/register', async (req, res) => {
+router.post('/user/register', async (req, res) => {
     User.findOne({phone: req.body.user.phone}, async (err, user) => {
         if(!user){
             user = new User(req.body.user);
@@ -33,7 +33,7 @@ router.post('/register', async (req, res) => {
 });
 
 //Verify the sms code and send the unique token that must be saved if it is correct
-router.post('/verify', async (req, res) => {
+router.post('/user/verify', async (req, res) => {
 
     User.findOne({phone: req.body.user.phone}, async (err, user) => {
         if(err){
@@ -56,11 +56,12 @@ router.post('/verify', async (req, res) => {
     });
 });
 
-router.post('/logTest', auth, async (req, res) => {
+router.post('/user/logTest', auth, async (req, res) => {
     res.status(200).send("All ok");
 });
 
-router.post('/logout', auth, async (req, res) => {
+//TODO: acabar logout
+router.post('/user/logout', auth, async (req, res) => {
     try {
         req.user.tokens = req.user.tokens.filter((token) => {
             return token.token !== req.token;
@@ -73,10 +74,45 @@ router.post('/logout', auth, async (req, res) => {
     }
 });
 
-router.post('/removeAllUsers', async (req, res) => {
-    await User.deleteMany({});
-    res.status(200).send("Removed all users");
-})
+router.delete('/user/removeAll', auth, async (req, res) => {
+
+    if(req.user.isAdmin) {
+        await User.deleteMany({});
+        res.status(200).send("Removed all users");
+    } else {
+        res.status(400).send("You have no permission");
+    }
+    
+});
+
+router.delete('/user/:id', auth, async (req, res) => {
+    User.findById(mongoose.Types.ObjectId(req.params.id), (err, user) => {
+        if(err){
+            res.status(400).send(err);
+            return;
+        }
+        res.status(200).send(`User with id ${req.params.id} and phone ${user.phone} removed`);
+    });
+
+});
+
+router.patch('/user/:id', auth, async (req, res) => {
+    var updateObject = req.body.user;
+    if(req.user.isAdmin){
+        User.findOneAndUpdate({_id: mongoose.Types.ObjectId(req.params.id)}, {$set: updateObject}, {new: true}, (err, user) => {
+            if (err){
+                res.status(400).send(err);
+                return;
+            }
+            if(!user){
+                res.status(400).send('No user found');
+            }
+            res.status(200).send(user);        
+        });
+    }else{
+        res.status(400).send('You are not an admin!');
+    }
+});
 
 module.exports = router;
